Split Mercator into Mercator/Parley
This patch splits what is currently `Mercator.swift` into
`Parley.swift` for all of the Charter, Tradewinds stuff, and
`Mercator.swift/` for all of the Mercator (app/UI) stuff.

Note: to facilitate this working, we're currently bringing in the stuff
required by Charter that _should probably_ not be in there, i.e.
`Prelude` and `Portfolios`.

Preferably we can remove this dependency entirely from `Parley.swift`
(and move these back into Mercator.swift), or, if need be, find a way to
share them. Overall, this is meant to reduce the amount of changes that
would require us to bump the `.mercator-version`, e.g. for changes
simply to the UI.